requirements.txt
MLB-StatsAPI>=1.7.0
pybaseball>=2.2.7
pandas>=2.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
python-dotenv>=1.0.0
tabulate>=0.9.0
colorama>=0.4.6
anthropic>=0.40.0
